@@ -77,16 +77,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.11",
+    "@angular/animations": "^14.0.6",
     "@angular/cdk": "^13.1.1",
-    "@angular/common": "^13.3.11",
-    "@angular/compiler": "^13.3.11",
-    "@angular/core": "^13.3.11",
-    "@angular/forms": "^13.3.11",
+    "@angular/common": "^14.0.6",
+    "@angular/compiler": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
+    "@angular/platform-browser": "^14.0.6",
+    "@angular/platform-browser-dynamic": "^14.0.6",
+    "@angular/router": "^14.0.6",
     "@azure/msal-angular": "^2.0.1",
     "@azure/msal-browser": "^2.15.0",
     "@ionic-native/core": "^5.10.0",
@@ -133,14 +133,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.8",
+    "@angular-devkit/build-angular": "^14.0.6",
     "@angular-eslint/builder": "14.0.2",
     "@angular-eslint/eslint-plugin": "14.0.2",
     "@angular-eslint/eslint-plugin-template": "14.0.2",
     "@angular-eslint/schematics": "14.0.2",
     "@angular-eslint/template-parser": "14.0.2",
-    "@angular/cli": "^13.3.8",
-    "@angular/compiler-cli": "^13.3.11",
+    "@angular/cli": "^14.0.6",
+    "@angular/compiler-cli": "^14.0.6",
     "@babel/core": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
     "@compodoc/compodoc": "^1.1.15",
@@ -177,7 +177,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.0.3",
     "npm-run-all": "^4.1.5",
     "protractor": "~7.0.0",
     "sass": "^1.42.1",
