@@ -25,7 +25,7 @@ import { FontType, GeometryType } from '../shared/draw.enum';
 import { IgoMap } from '../../map/shared/map';
 import { BehaviorSubject, Subscription } from 'rxjs';
 import { Draw, FeatureWithDraw } from '../shared/draw.interface';
-import { FormGroup, FormBuilder } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder } from '@angular/forms';
 import { VectorSourceEvent as OlVectorSourceEvent } from 'ol/source/Vector';
 import { VectorLayer } from '../../layer/shared/layers/vector-layer';
 import { FeatureDataSource } from '../../datasource/shared/datasources/feature-datasource';
@@ -109,13 +109,13 @@ export class DrawComponent implements OnInit, OnDestroy {
   public fontSizeForm: string;
   public fontStyleForm: string;
   public position: string = 'bottom';
-  public form: FormGroup;
+  public form: UntypedFormGroup;
   public icons: Array<string>;
   public icon: string;
 
   constructor(
     private languageService: LanguageService,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private drawStyleService: DrawStyleService,
     private dialog: MatDialog,
     private drawIconService: DrawIconService
